Struct hawk::Header

pub struct Header {
    pub id: Option<String>,
    pub ts: Option<Timespec>,
    pub nonce: Option<String>,
    pub mac: Option<Mac>,
    pub ext: Option<String>,
    pub hash: Option<Vec<u8>>,
    pub app: Option<String>,
    pub dlg: Option<String>,
}

Representation of a Hawk Authorization header value (the part following "Hawk ").

Headers can be derived froms trings using the FromStr trait, and formatted into a string using the fmt_header method.

All fields are optional, although for specific purposes some fields must be present.

Fields
id: Option<String> ts: Option<Timespec> nonce: Option<String> mac: Option<Mac> ext: Option<String> hash: Option<Vec<u8>> app: Option<String> dlg: Option<String>
Methods

impl Header

pub fn new<S>(
id: Option<S>,
ts: Option<Timespec>,
nonce: Option<S>,
mac: Option<Mac>,
ext: Option<S>,
hash: Option<Vec<u8>>,
app: Option<S>,
dlg: Option<S>
) -> Result<Header> where
S: Into<String>,

Create a new Header with the full set of Hawk fields.

This is a low-level funtion. Headers are more often created from Request or Responses.

Note that none of the string-formatted header components can contain the character \".

pub fn fmt_header(&self, f: &mut Formatter) -> Result

Format the header for transmission in an Authorization header, omitting the "Hawk " prefix.