Struct hawk::RequestBuilder

pub struct RequestBuilder<'a>(_);

Methods

impl<'a> RequestBuilder<'a>

pub fn new(method: &'a str, host: &'a str, port: u16, path: &'a str) -> Self

Create a new request with the given method, host, port, and path.

pub fn from_url(method: &'a str, url: &'a Url) -> Result<Self>

Create a new request with the host, port, and path determined from the URL.

pub fn method(self, method: &'a str) -> Self

Set the request method. This should be a capitalized string.

pub fn path(self, path: &'a str) -> Self

Set the URL path for the request.

pub fn host(self, host: &'a str) -> Self

Set the URL hostname for the request

pub fn port(self, port: u16) -> Self

Set the URL port for the request

pub fn url(self, url: &'a Url) -> Result<Self>

Set the hostname, port, and path for the request, from a string URL.

pub fn hash<H: Into<Option<&'a [u8]>>>(self, hash: H) -> Self

Set the content hash for the request

pub fn ext<S: Into<Option<&'a str>>>(self, ext: S) -> Self

Set the ext Hawk property for the request

pub fn app<S: Into<Option<&'a str>>>(self, app: S) -> Self

Set the app Hawk property for the request

pub fn dlg<S: Into<Option<&'a str>>>(self, dlg: S) -> Self

Set the dlg Hawk property for the request

pub fn request(self) -> Request<'a>

Get the request from this builder