Struct hawk::ResponseBuilder

pub struct ResponseBuilder<'a>(_);

Methods

impl<'a> ResponseBuilder<'a>

pub fn from_request_header(
req_header: &'a Header,
method: &'a str,
host: &'a str,
port: u16,
path: &'a str
) -> Self

Generate a new Response from a request header.

This is more commonly accessed through Request::make_response.

pub fn hash<H: Into<Option<&'a [u8]>>>(self, hash: H) -> Self

Set the content hash for the response.

This should always be calculated from the response payload, not copied from a header.

pub fn ext<S: Into<Option<&'a str>>>(self, ext: S) -> Self

Set the ext Hawk property for the response.

This need only be set on the server; it is ignored in validating responses on the client.

pub fn response(self) -> Response<'a>

Get the response from this builder